backend/pyproject.toml
[tool.poetry]
name = "isi-backend"
version = "0.1.0"
description = ""
authors = ["Rafał Łazicki <rafal.lazicki99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
